+++ packages/generator-ts-project/src/Project/VSCode/TSProjectLaunchSettingsProcessor.ts
@@ -56,7 +56,11 @@ export class TSProjectLaunchSettingsProcessor<TSettings extends ITSProjectSettin
      */
     protected override async FilterDebugConfig(debugConfig: DebugConfiguration): Promise<boolean>
     {
-        return (debugConfig.program ?? "").includes(this.GetWorkspaceFolderDirective("TSProjectGenerator")) &&
+        let workspaceDirective = this.GetWorkspaceFolderDirective("TSProjectGenerator");
+
+        return (
+            (debugConfig.args as string[])?.[0]?.includes(workspaceDirective) ||
+            (debugConfig.cwd as string)?.includes(workspaceDirective)) &&
             !(normalize(debugConfig.program ?? "").toLowerCase().endsWith(join("node_modules", "yo", "lib", "cli.js")));
     }